pyproject.toml
[project]
name = "imas-standard-names"
dynamic = ["version"]
description = "A collection of Standard Names used in the Fusion Conventions and logic for creating a static website for documentation."
authors = [
  { name = "Maarten Sebregts", email = "msebregts@ignitioncomputing.com" },
  { name = "Simon McIntosh", email = "simon.mcintosh@iter.org" },
]
license = { text = "CC BY-ND 4.0" }
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
  "strictyaml (>=1.7.3,<2.0.0)",
  "pydantic (>=2.10.6,<3.0.0)",
  "click (>=8.1.8,<9.0.0)",
  "pyyaml (>=6.0.2,<7.0.0)",
  "pint (>=0.24.4,<0.25.0)",
  "pandas (>=2.2.3,<3.0.0)",
  "tabulate (>=0.9.0,<0.10.0)",
  "requests (>=2.32.3,<3.0.0)",
  "markdown (>=3.8,<4.0)",
  "imas-mcp>=2.4",
  "textual>=6.1.0",
  "pydantic-ai>=1.0.8",
  "pydantic-evals>=0.0.7",
  "pydantic-graph>=0.0.2",
  "proselint (>=0.14.0,<0.15.0)",
  "spacy (>=3.8.0,<4.0.0)",
  "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
]
keywords = [
  "fusion",
  "standard names",
  "IMAS",
  "ITER",
  "fusion conventions",
  "fusion data",
  "fusion research",
]
[project.urls]
"Documentation" = "https://iterorganization.github.io/IMAS-Standard-Names/"

[project.scripts]
has_standardname = "imas_standard_names.issues.cli:has_standardname"
update_standardnames = "imas_standard_names.issues.cli:update_standardnames"
get_standardname = "imas_standard_names.issues.cli:get_standardname"
is_genericname = "imas_standard_names.issues.cli:is_genericname"
update_links = "imas_standard_names.issues.cli:update_links"
validate_catalog = "imas_standard_names.validation.cli:validate_catalog_cli"
standard-names = "imas_standard_names.cli:standard_names"
standard-names-mcp = "imas_standard_names.cli.server:main"
build-grammar = "imas_standard_names.grammar_codegen.generate:main"

[project.optional-dependencies]
agent = [
  "dotenv (>=0.9.9,<0.10.0)",
  "fastmcp (>=2.2.6,<3.0.0)",
  "pydantic-ai (>=1.0.0,<2.0.0)",
  "imas-python (>=2.0.0,<3.0.0); python_version < '3.13'",
]
test = [
  "pytest (>=8.3.4,<9.0.0)",
  "pytest-cov (>=4.1.0,<5.0.0)",
  "pytest-html (>=4.1.1,<5.0.0)",
  "ruff (>=0.9.8,<1.0.0)",
]
docs = [
  "mike (>=2.1.3,<3.0.0)",
  "mkdocs (>=1.6.1,<2.0.0)",
  "mkdocs-macros-plugin (>=1.0.4,<2.0.0)",
  "mkdocs-material (>=9.6.5,<10.0.0)",
  "mkdocs-data-plugin (>=0.2.0,<0.3.0)",
  "mkdocs-table-reader-plugin (>=3.1.0,<4.0.0)",
  "mkdocs-include-markdown-plugin (>=7.0.0,<8.0.0)",
]
[build-system]
requires = ["hatchling>=1.25.0", "hatch-vcs>=0.4.0", "pyyaml>=6.0.2", "pydantic>=2.10.6"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.hooks.custom]
path = "hatch_build_hooks.py"

[dependency-groups]
dev = [
  "pre-commit>=4.1.0,<5.0.0",
  "ruff>=0.11.10,<1.0.0",
  "pytest>=8.3.4,<9.0.0",
  "pytest-cov>=4.1.0,<5.0.0",
  "pytest-html>=4.1.1,<5.0.0",
  "ipykernel>=6.29.5,<7.0.0",
  "pandas-stubs>=2.2.3.250308,<3.0.0",
  "types-pyyaml>=6.0.12.20241230,<7.0.0",
  "logfire>=3.14.0,<4.0.0",
  "mcp-cli>=0.1.0,<1.0.0",
  "textual-dev>=1.7.0",
]

[[tool.mypy.overrides]]
module = ["strictyaml"]
ignore_missing_imports = true

[tool.coverage.run]
omit = ['__init__.py']

[tool.pytest.ini_options]
# No special anyio config needed - handled in conftest.py

[tool.ruff]
target-version = "py312"
line-length = 88
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "PLC0415",  # import-outside-top-level
]
ignore = [
    "C901",
    "E501", 
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011","F401","F811"]

[tool.ruff.lint.isort]
known-first-party = ["imas_standard_names"]
combine-as-imports = true
